Smarter PlaceNode.Country(). Fixes #143 (#147)
For places that do not follow the strict four jurisdictional entities the country can still be extracted by referring to a list of known countries and comparing them with the suffix of the place name.

It can work around some basic puntuation, capitalization and whitespace. However, it is not as full proof as it could be, yet.