Remove Document.Nodes
Document.Nodes is no longer publically accessible and added a NewDocumentWithNodes() constructor.

Instead there are new methods Nodes() and AddNode() on the Document. These are required to make sure the proper cache mechanisms now and in the future can be triggered when the document changes.

It was stated before that you should never instantiate Document manually. However, all of the tests that did this have now been updated to use one of the constructors.