Adding Age type and calculations. Fixes #9
This is a surprisingly large change. It introduces several new types and a lot of new functionality centered around the main new type, Age.

Age can be use used to describe a real or theoretical age of an individual at the time of their death or another event. The Age containts other attributes such as whether it is an estimation, handles date ranges and calculated metrics about whether they would be alive during the date range.

Most of the new functionality can be seen through IndividualNode.Age() and IndividualNode.AgeOf().

The other changes worth noting are:

- DateNode.IsValid returns true only when the node is not nil and the start and end date are non-zero.
- DateNode.IsExact and Date.IsExact will return true if the date range represents a single day with an exact constraint.
- Fixed a bug where DateNodes.Minimum and DateNodes.Maximum used an average which wasn't sensible with date ranges. Now it will consider the values as a range rather than a point in time.
- DateNodes.StripZero returns a new slice that only contains dates that are not zero (basically, valid and usable).
- NewDateWithTime creates a new Date with the provided time.Time.
- NewDateWithNow creates a two Dates that represents the the start and end of the current day.