Introducing gedcomq
Package gedcomq is a command line tool and query language for GEDCOM files. It is heavily inspired by jq, in name and syntax. The basic syntax of the tool is:

  gedcomq -gedcom file.ged '.Individuals | .Name'

The functions are automatically discovered through the gedom package.

This initial version is quite rudimentary and does not handle errors gracefully. However, it does demonstrate what the most basic queries look like.