gedcomdiff: Added -prefer-pointer-above
Controls if two individuals should be considered a match by their pointer value.

The default value is 0.733 which means that the individuals will be considered a match if they share the same pointer and hit the same default minimum similarity.

A value of 1.0 would have to be a perfect match to be considered equal on their pointer, this is the same as disabling the feature.

A value of 0.0 would mean that it always trusts the pointer match, even if the individuals are nothing alike.

This option makes sense when you are comparing documents that have come from the same base and retained the pointers between individuals of the existing data.