dates: added phrases
A phrase is any statement offered as a date when the year is not recognizable to a date parser, but which gives information about when an event occurred. The date phrase is enclosed in matching parentheses.

Phrases can only be compared to themselves and they must be the exact same value to be considered equal.