_UID: Unofficial unique ID (#218)
The _UID tag is what I call a Common GEDCOM Extension; a GEDCOM extension that's common to many different applications. The _UID tag is a legal GEDCOM extension; it starts with an underscore, like any vendor-defined extension should. However, it would be wrong to think of the _UID tag as a vendor- or product-specific tag. The _UID tag is not specific to any particular product or vendor. It is a common GEDCOM extension, supported by quite a few different products from different vendors.