Refactoring DateNode to DateRange (#234)
Most of the date range functionality was in the DateNode which means it could not be used in other places we need to represent date ranges. There is now a DateRange type where most of these methods have been moved to.

Also, more comprehensive comparisons can be made like Before, After and range comparison constants.