Warning: Unparsable date (#237)
Warnings are a way to surface data that may be wrong for a variety of reasons. Some examples may be that dates cannot be understood, children are born before parents, nodes to adhere to GEDCOM standards, etc.

Nodes can now implement a Warner interface which returns a slice of Warning entities. Document also implements Warner to aggregate all warnings.

This patch starts with a single type of warning, when dates cannot be parsed. More kinds of warnings will be added in time.