Decoder: Allow mulitline GEDCOM values
It is not valid for GEDCOM values to contain new lines or carriage returns. However, some application dump data without correctly using the CONT tags.

Strictly speaking we should bail out with an error but there are too many cases that are difficult to clean up for consumers so we offer and option to permit it.

When enabled any line than cannot be parsed will be considered an extension of the previous line (including the new line character).