Added IncorrectEventOrderWarning
IncorrectEventOrderWarning produces warnings where events for an individual are in an incorrect order. For example a burial before a death event, or a baptism before a birth.

This patch also creates a SimpleWarning so there is less duplicate code between warnings and a few helper methods for events on IndividualNode.