HTMLFormatter for gedcomq (#253)
The q engine now supports outputting in HTML with "-format html". Internal types can implement Component like any other html package component to be rendered in such a way. If the object does not implement Component then its pretty JSON will be output as plain fixed-width text.

To make this happen I've had to split the html package into html and html/core where core contains only components that do not need to reference the gedcom package, such as a HTML table. This is to get around a cyclic import.

Warnings in an entire document can now be rendered with:  gedcomq -gedcom file.ged -format html '.Warnings'

The QMarshaller has been removed. This was a stopgap so that the JSON outputted from warnings wasn't so verbose, but now that's not needed with the HTML formatter.

The core.Component (previously html.Component) renames the WriteTo() to WriteHTMLTo(). This makes a lot more sense when writing compatible components outside of the html package.