Decoder: AllowMultiLine mangles some nodes (#262)
Since AllowMultiLine will allow blank lines to continue there is a bug where this might lead to node values having a traing new line.

For these cases the error cascades down to node that need to decode the value (such as a date) and the added newline break the regexps.

Now new lines can still exist in multiline values, but the value itself cannot end with a new line. This is a good balance been compatibility and consistence.