gedcom2html: Faster and customizable rendering
The code to publish a website has now been moved into the html package as Publisher. This means that the gedcom2html is much lighter.

The Publisher can now be customised to write to the file system, or any other storage mechanism and it can now render the website in parallel making it much faster. This can be configured with the new "-jobs" option.

The "-checksum" option has been removed from the gedcom2html because it was already problematic at best and not needed now that the rendering process is much faster.