decoder: Added AllowInvalidIndents
AllowInvalidIndents allows a child node to have an indent greater than +1 of the parent. AllowInvalidIndents is disabled by default because if this happens the GEDCOM file is broken in some possibly serious way and certainly not a valid GEDCOM file.

The biggest problem with having the indents wrongly aligned is that nodes that are expected to be a certain depth (such as NPFX inside a NAME) will probably break or interfere with a traversal algorithm that is not expecting the node to be there/at that level.

Another important thing to note is that the incorrect indent level will not be retained when writing the Document back to a GEDCOM.