Fix missing "Age" in publish events
The age of the individual at that event would not show if the death date
was missing. This is because it would consider the event always after
their death, which is normally hidden from publish events.

Also fixed a bug where the places map (also in publish) was global.

Fixes #295